// KLMPNHomeStay/ClientApp/src/app/model/secured/package.ts
export class Package {
  PName: string;
  DName: string;
  Day: string;
  Night: string;
  Subject: string;
  Description: string;
  Cost: string;
  ContactPerson: string;
  ContactEmail: string;
  ContactMobile: string;
  Image1: string;
  Image2: string;
  Image3: string;
  Image4: string;
  Image5: string;
  PackagePdf: string;
  roomFacility: any;
  CreatedBy: string;
}